Add support for a custom tag prefix
This branch adds support for specifiying a tag prefix before the version
number. It turns out that this was actually pretty simple to add. We
simply interpolate the passed-in prefix to the regular expression for
detecting tags, and then strip the prefix when generating the version
number.

Currently, the prefix is interpreted as a fixed string; if it contains
regular expression characters, we escape them before adding it to the
pattern that's matched to detect tags that should generate releases.
It's possible that maybe we'd actually _prefer_ to allow interpreting it
as a pattern, so I'd be happy to change that if so.